requirements.txt
asgiref==3.8.1
babel==2.17.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==3.1.14
django-countries==7.6.1
django-extra-views==0.13.0
django-filter==2.4.0
django-haystack==3.2.1
django-oscar==3.0.1
django-oscar-api==3.1.2
django-phonenumber-field==3.0.1
django-tables2==2.3.4
django-tenant-schemas==1.12.0
django-tenants==3.7.0
django-treebeard==4.4
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
factory-boy==2.12.0
Faker==36.1.1
graphene==3.4.3
graphene-directives==0.4.6
idna==3.10
ordered-set==4.1.0
packaging==24.2
phonenumbers==8.13.55
pillow==11.1.0
promise==2.3
psycopg2-binary==2.9.10
purl==1.6
PyJWT==2.10.1
pysolr==3.10.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyuca==1.2
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sorl-thumbnail==12.9.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
